update: removing duplicate codes!
