+++ dags/hivemind_etl.py
@@ -29,8 +29,7 @@ from hivemind_etl_helpers.discord_mongo_summary_etl import process_discord_summa
 from hivemind_etl_helpers.discord_mongo_vector_store_etl import (
     process_discord_guild_mongo,
 )
-from hivemind_etl_helpers.src.utils.mongo import MongoSingleton
-
+from hivemind_etl_helpers.src.utils.get_mongo_discord_communities import get_all_discord_communities
 
 # def setup_phonix():
 #     _ = px.launch_app()
@@ -49,18 +48,11 @@ with DAG(
 ) as dag:
 
     @task
-    def get_all_discord_communities() -> list[str]:
+    def get_discord_communities() -> list[str]:
         """
         Getting all communities having discord from database
         """
-        mongo = MongoSingleton.get_instance()
-        communities = (
-            mongo.client["Core"]["platforms"]
-            .find({"name": "discord"})
-            .distinct("community")
-        )
-        # getting the str instead of ObjectId
-        communities = [str(comm) for comm in communities]
+        communities = get_all_discord_communities()
         return communities
 
     @task
@@ -81,12 +73,7 @@ with DAG(dag_id="discord_summary_vector_store", start_date=datetime(2023, 1, 1))
         """
         Getting all communities having discord from database
         """
-        mongo = MongoSingleton.get_instance()
-        communities = (
-            mongo.client["Core"]["platforms"]
-            .find({"name": "discord"})
-            .distinct("community")
-        )
+        communities = get_all_discord_communities()
         return communities
 
     @task
